nl80211/cfg80211: add match filtering for sched_scan
Introduce filtering for scheduled scans to reduce the number of
unnecessary results (which cause useless wake-ups).

Add a new nested attribute where sets of parameters to be matched can
be passed when starting a scheduled scan.  Only scan results that
match any of the sets will be returned.

At this point, the set consists of a single parameter, an SSID.  This
can be easily extended in the future to support more complex matches.

Signed-off-by: Luciano Coelho <coelho@xxxxxx>
---

This replaces "nl80211/cfg80211: add ssid filtering for sched_scan".
The commit has been renamed to better reflect the new logic.

Changes in v2:
	* changed the concept to use match sets instead of different
	types of filters (as discussed in v1);